Move tarball logic back to lifecycle
